Use authService to manage ticket usage
